+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Derive random IV if not given based on key, not in default value to function arguments
+
 ## [v1.0.3] - 2024-12-16
 
 ### Added